softwares/cursos/requirements.txt
# Flask==2.3.3
# Flask-SQLAlchemy==3.1.1
# Flask-Migrate==4.0.4
# Flask-WTF==1.1.1
# Flask-Login==0.6.2
# Flask-JWT-Extended==4.5.2
# Flask-CORS==4.0.0
# Flask-Mail==0.9.1
# Flask-Bcrypt==1.0.1
# Flask-Caching==2.0.2
# psycopg2-binary==2.9.7
# SQLAlchemy==2.0.20
# WTForms==3.0.1
# email-validator==2.0.0
# python-dotenv==1.0.0
# Pillow==10.0.0
# google-api-python-client==2.97.0
# google-auth==2.22.0
# google-auth-oauthlib==1.0.0
# google-auth-httplib2==0.1.0
# google-generativeai==0.3.1
# requests==2.31.0
# pyopenssl==23.2.0
# passlib==1.7.4
# pytest==7.4.0
# pytest-flask==1.2.0
# pytest-cov==4.1.0
# flake8==6.1.0
# black==23.7.0
# gunicorn==21.2.0
# python-slugify==8.0.1
# Werkzeug==2.3.7
# redis==5.0.0
# ipython==8.14.0
# pyyaml==6.0.1
# tenacity==8.2.3

Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login
Flask-JWT-Extended
Flask-CORS
Flask-Mail
Flask-Bcrypt
Flask-Caching
psycopg2-binary
SQLAlchemy
WTForms
email-validator
python-dotenv
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
requests
pyopenssl
passlib
pytest
pytest-flask
pytest-cov
flake8
black
gunicorn
python-slugify
Werkzeug
redis
ipython
pyyaml
tenacity